Prefer methods over signals
When a method with the same name is found, then 'Event' is added to the signal's name. It seems that in most cases the method name should be preferred over the signal name. Also emit a warning tag, when a method is renamed, because these events should be renamed using the metafile manually.